# Optimize UnicodeICU localeCompare